Add files via upload
Backup script that is used as a form of regularly backing up items. It creates a log file that dates it so you can keep track of when the backup occurs and logs what gets backed up and what receives an error, etc. 

$filename is the name of the log file
$filedate adds the date to the file name
$folderdate copies the file to a folder with the year/month
#options2 can be used with /xd to exclude entire directories
or /xf to exclude a file 
(/XD exclude-folder-1) (/XF filename-1.extension)

Place and run this file in the folder you want your backup
logs to be created and stored.